normalyser/requirements.txt
langchain==0.3.15
langchain-openai==0.3.2
python-dotenv==1.0.0
grpcio==1.70.0
grpcio-tools==1.70.0
